The ARM MOVES!!!!!
The arm has a full range of motion. The vex limit switch works and it homes (we still need to place the switch on the robot, but manually it works).

The arm has 3 preset positions, but at the moment, only high and stow work without stalling. The arm constants still need to be tuned.

Summary, the arm works and all we need to do is tune it now.