Arm and Intake Move correctly
The arm and Intake have PID and code that allows them to move to positions upon button press. The Intake rollers on both the arm and ground intake have been synced up so that the bumpers control both the ground intake and the arm intake.

The Intake needs to be zeroed before being turned on or being built in order to prevent the gears from skipping and tearing off teeth.